// src/interfaces/page.interface.ts
export enum TopLevelCategory {
  Courses,
  Services,
  Books,
  Products,
}

export interface TopPageModal {
  _id: string
  tags: string[]
  secondCategory: string
  alias: string
  title: string
  category: string
  seoText: string
  tagsTitle: string
  metaTitle: string
  metaDescription: string
  firstCategory: TopLevelCategory
  advantages: TopPageAdvantage[]
  createdAt: string
  updatedAt: string
  __v: number
  hh: HhData
  categoryOn: string
}

export interface TopPageAdvantage {
  title: string
  description: string
  _id: string
}

export interface HhData {
  count: number
  juniorSalary: number
  middleSalary: number
  seniorSalary: number
  updatedAt: string
  _id: string
}